cleaning + making function thats makes request to usos api
